requirements.txt
# --- Core Web ---
Flask==3.0.0
gunicorn==21.2.0
Werkzeug==3.0.1
Jinja2==3.1.3
itsdangerous==2.1.2
click==8.1.7

# --- Telegram Bot ---
python-telegram-bot==21.6
httpx==0.27.0

# --- Env & Config ---
python-dotenv==1.0.1

# --- Auth / Security ---
Flask-Login==0.6.3
Flask-WTF==1.2.1
email-validator==2.1.1

# --- Database ---
Flask-SQLAlchemy==3.1.1
SQLAlchemy==2.0.25
psycopg2-binary==2.9.9   # PostgreSQL (Render native)

# --- Forms & UI ---
WTForms==3.1.2
Flask-Bootstrap==3.3.7.1

# --- Utils ---
requests==2.32.3
pytz==2024.1

# --- Excel ---
openpyxl==3.1.2
